updated for loop interators

+++ TriangleRenderer/Scene.cpp
@@ -25,7 +25,7 @@ void Scene::updateObjects(float DeltaTime)
 	//camPos = objects[2].GetPosition() + glm::vec3(0.0f, 2.0f, 10.0f);
 
 
-	for (int i = 0; i < objects.size(); i++)
+	for (size_t i = 0; i < objects.size(); i++)
 	{
 		objects[i].UpdatePhysics(DeltaTime, objects, i);
 	}
